### 2022-12-12 #5005 @editorialbot
[PRE REVIEW]: Trash AI: A Web GUI for Serverless Computer Vision Analysis of Images of Trash
**Submitting author:** @wincowgerDEV (<a href="http://orcid.org/0000-0001-9226-3104">Win Cowger</a>)
**Repository:** https://github.com/code4sac/trash-ai
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
