### 2022-12-17 #5005 @wincowgerDEV
Thanks so much for following up on this. No worries at all about
waitlisting. We anticipated that especially with the holidays coming up.
Hope you're able to get some vacation time. I will check in with the JS
team about how much of the JS is written in typescript vs vue and get back
to you. I will also follow up on the aspects listed previously after the
holidays in early January.
