### 2022-12-31 #5005 @wincowgerDEV
> @wincowgerDEV – thanks for your submission to JOSS. Can you help me understand where the code contributions are for this project? It looks to be a Python-based project but I'm trying to understand if significant aspects of the codebase are indeed in TypeScript.
>
> Any pointers you could give us here would be very welcome. Also as we're managing a backlog of submissions in this area I think it's appropriate to add the `waitlisted` label for now.

@arfon I got confirmation from the JS team that the JS is written in Typescript. They think that a reviewer familiar with JS could review the Typescript without much challenge.
