### 2022-12-31 #5005 @wincowgerDEV
updated DOIs https://github.com/code4sac/trash-ai/commit/52380b0a1915526db7c79b5cb44daad02b525a4c
