### 2023-01-12 a6f4242
deleted about.Vue (#90)
* Local Deployment (#86)

* Testing fixes in issue 85

* changed platform in dockerfile

* add wsl2 instructions

* add windows docker install instructions

* Update localdev.md

* reorganization

* Update localdev.md

* Update localdev.md

* cleanup

* Update localdev.md

* Update localdev.md

* fix error that was preventing initial install.

Co-authored-by: Steven Hollingsworth <shollingsworth@barracuda.com>

* updated openid thumbprint for github (#87)

* Manuscript (#69)

* my workflow

* example templates for manuscript.

* manuscript info

* update citations

* cl
[…]
