+++ paper.bib
@@ -1,0 +1,196 @@
+
+@software{Jocher:2020,
+  author       = {Glenn Jocher and
+                  Alex Stoken and
+                  Jirka Borovec and
+                  NanoCode012 and
+                  ChristopherSTAN and
+                  Liu Changyu and
+                  Laughing and
+                  tkianai and
+                  Adam Hogan and
+                  lorenzomammana and
+                  yxNONG and
+                  AlexWang1900 and
+                  Laurentiu Diaconu and
+                  Marc and
+                  wanghaoyang0106 and
+                  ml5ah and
+                  Doug and
+                  Francisco Ingham and
+                  Frederik and
+                  Guilhen and
+                  Hatovix and
+                  Jake Poznanski and
+                  Jiacong Fang and
+                  Lijun Yu 于力军 and
+                  changyu98 and
+                  Mingyu Wang and
+                  Naman Gupta and
+                  Osama Akhtar and
+                  PetrDvoracek and
+                  Prashant Rai},
+  title        = {{ultralytics/yolov5: v3.1 - Bug Fixes and 
+                   Performance Improvements}},
+  month        = oct,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v3.1},
+  doi          = {10.5281/zenodo.4154370},
+  url          = {https://doi.org/10.5281/zenodo.4154370}
+}
+
+@TECHREPORT{Moore:2020,
+  title       = "California Trash Monitoring Methods and Assessments Playbook",
+  author      = "Moore, Shelly and Hale, Tony and Weisberg, Stephen B and
+                 Flores, Lorenzo and Kauhanen, Pete",
+  institution = "San Francisco Estuary Institute",
+  year        =  2020
+}
+
+@ARTICLE{Hapich:2022,
+  title    = "Trash Taxonomy Tool: harmonizing classification systems used to
+              describe trash in environments",
+  author   = "Hapich, Hannah and Cowger, Win and Gray, Andrew and Tangri, Neil
+              and Hale, Tony and Magdy, Amr and Vermilye, Antoinette and Yu,
+              Walter and Ayres, Dick and Moore, Charles and Vermilye, John and
+              Singh, Samiksha and Haiman, Aaron N K and Youngblood, Kathryn and
+              Kang, Yunfan and McCauley, Margaret and Lok, Trevor and Moore,
+              Shelly and Baggs, Eric and Lippiatt, Sherry and Kohler, Peter and
+              Conley, Gary and Taing, Janna and Mock, Jeremiah",
+  abstract = "Despite global efforts to monitor, mitigate against, and prevent
+              trash (mismanaged solid waste) pollution, no harmonized trash
+              typology system has been widely adopted worldwide. This impedes
+              the merging of datasets and comparative analyses. We addressed
+              this problem by 1) assessing the state of trash typology and
+              comparability, 2) developing a standardized and harmonized
+              framework of relational tables and tools, and 3) informing
+              practitioners about challenges and potential solutions. We
+              analyzed 68 trash survey lists to assess similarities and
+              differences in classification. We created comprehensive
+              harmonized hierarchical tables and alias tables for item and
+              material classes. On average, the 68 survey lists had 20.8\% of
+              item classes in common and 29.9\% of material classes in common.
+              Multiple correspondence analysis showed that the 68 surveys were
+              not significantly different regarding organization type,
+              ecosystem focus, or substrate focus. We built the Trash Taxonomy
+              Tool (TTT) web-based application with query features and open
+              access at openanalysis.org/trashtaxonomy. The TTT can be applied
+              to improve, create, and compare trash surveys, and provides
[…]

[…]
