### 2023-01-22 #5005 @arfon
@wincowgerDEV – apologies for the delay here. I'm still struggling to understand exactly what this submission is composed of. It looks to be:

- Some kind of pre-trained ML model for detecting items of trash in images.
- A web-ui for driving the model (on serverless compute)
- Some code for deploying the infrastructure

Is that about right? If so, I'm not sure this is eligible for review in JOSS as we don't review ML models here and generally review underlying libraries/codebases that are designed for [maintainable extension](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) rather than user-interfaces driving them (which is where the majority of the code in this submission seems to be focused).

To be clear, I'm not entirely convicted about my feedback here but as I conduct my initial assessment here this is where my head is at.
