### 2023-01-22 #5005 @wincowgerDEV
@arfon

Thanks for getting back to me and no worries about the delay. You are
correct with your three points about the major aspects of our software. I
will add though that the UI and backend itself is a substantial codebase
development, the deployment code is too but is a smaller component of this
work and we are asking for review of those, not the ML model.  The reason I
decided to submit this software to JOSS was inspired by a review I did for
JOSS. You can see the review here:
https://github.com/openjournals/joss-reviews/issues/4038

The software was very similar to ours in that it primarily made adoption of
ML and AI by linking existing UI packages together in a new and useful way
that advances AI and ML practitioners ability to conduct their science.
Doing computer vision without the aid of UIs is extremely challenging so I
anticipate this area of statistics will need many open source UI software
packages to move forward.

How I understood the maintainable extension requirement during that review
was that the code needed to demonstrate enough uniqueness to stand alone
and that their needed to be clear enough guidance to promote contribution
and maintenance by others in the future. Our software has these features.
There are clear contribution guidelines in the repo and a team composed of
software developers around the USA affiliated with the volunteer group Code
for America developed the code with many hundreds of hours of effort.

Happy to continue the conversation and appreciate you willingness to
discuss.

Warm Regards,
Win
