### 2023-02-02 c61de3e
Bump simple-git from 3.5.0 to 3.16.0 in /backend (#95)
* Local Deployment (#86)

* Testing fixes in issue 85

* changed platform in dockerfile

* add wsl2 instructions

* add windows docker install instructions

* Update localdev.md

* reorganization

* Update localdev.md

* Update localdev.md

* cleanup

* Update localdev.md

* Update localdev.md

* fix error that was preventing initial install.

Co-authored-by: Steven Hollingsworth <shollingsworth@barracuda.com>

* updated openid thumbprint for github (#87)

* Manuscript (#69)

* my workflow

* example templates for manuscript.

* manuscript
[…]
