+++ localdev/Makefile
@@ -2,6 +2,7 @@ down:
 	docker compose down
 
 up: down
+
 	docker compose build
 	docker compose up --remove-orphans
 	make down
