### 2023-02-04 180523e
Local Deployment (#86)
* Testing fixes in issue 85

* changed platform in dockerfile

* add wsl2 instructions

* add windows docker install instructions

* Update localdev.md

* reorganization

* Update localdev.md

* Update localdev.md

* cleanup

* Update localdev.md

* Update localdev.md

* fix error that was preventing initial install.

Co-authored-by: Steven Hollingsworth <shollingsworth@barracuda.com>
