### 2023-02-05 #5005 @arfon
@luxaritas, @domna, @a2un, @MikeHughesKent :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is Trash AI: A Web GUI for Serverless Computer Vision Analysis of Images of Trash

Many thanks
Arfon
