### 2023-02-06 #5005 @domna
Hi @arfon ,

I would be happy to review the paper. It would be my first review for JOSS. Judging from the reviewer guidelines I feel confident to deliver a good review for the paper. It may just take a little longer since it's my first time and I just need to adapt to the processes.
