### 2023-02-13 code4sac/trash-ai#106 @wincowgerDEV
[Feature]: consolidate deployment scripts
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Description

Reviewer 1 says: Even with the current structure, I'll also note that the deployment scripts and such could probably be simplified and consolidated a bit - when taking a brief look at it, it felt like I had to jump around a bit to find all the different bits and pieces and had a lot of different calls between different tools (eg, why use makefiles vs just having a single docker-compose in the root and use docker-compose up?).

### Problem

Challenging deployment

### Proposed Solution

Make deployment scripts easier to follow.

### Alternatives Considered

NA
