### 2023-02-13 code4sac/trash-ai#107 @wincowgerDEV
[Feature]: Better deployment and use instructions
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Description

Reviewer 1: While the video walkthrough is great, it would probably be best to have a bit more written instruction on usage in the README/on the website. I think the setup documentation could also be improved a bit. The local setup seems to refer to specifics of WSL a lot where it could be more system-agnostic, and while the usage of GH workflows is novel, the AWS deployment instructions are a bit nonstandard in that regard plus seem to assume that you're deploying it from the repo itself/as a repo owner.

### Problem

Instructions are overly specific and not enough information on the website.

### Proposed Solution

System agnostic deployment instructions and more information on the website.

### Alternatives Considered

NA
