### 2023-02-13 code4sac/trash-ai#108 @wincowgerDEV
[Feature]: Instructions to verify behavior and Tests
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Description

There doesn't seem to be any automated tests. While the desired behavior here is relatively clear/trivial, there's no formal instructions to verify behavior.

### Problem

NO

### Proposed Solution

Need to add formal instructions to verify behavior and add clarity on the automated tests currently implemented.

### Alternatives Considered

NA
