### 2023-02-13 code4sac/trash-ai#109 @wincowgerDEV
[Bug]: References
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Project Version

_No response_

### Platform and OS Version

_No response_

### Existing Issues

_No response_

### What happened?

OK DOIs

- 10.5281/zenodo.4154370 is OK
- 10.48550/ARXIV.2003.06975 is OK

MISSING DOIs

- 10.1186/s43591-022-00035-1 may be a valid DOI for title: Trash Taxonomy Tool: harmonizing classification systems used to describe trash in environments
- 10.1186/s40965-018-0050-y may be a valid DOI for title: OpenLitterMap.com – Open Data on Plastic Pollution with Blockchain Rewards (Littercoin)

INVALID DOIs

- https://doi.org/10.1029/2019EA000960 is INVALID because of 'https://doi.org/' prefix
- https://doi.org/10.1016/j.wasman.2021.12.001 is INVALID because of 'https://doi.org/' prefix

### Steps to reproduce

Automated JOSS Bot.

### Expected behavior

DOI's to be valid.

### Attachments

_No response_

### Screenshots or Videos

_No response_

### Additional Information

_No response_
