### 2023-02-13 #5136 @wincowgerDEV
Hey @luxaritas, Thanks for the initial pass. I am creating some issues now based on these helpful comments and have responses below. Will update you when the issues are fixed.

It does not appear the submitting author (@wincowgerDEV) made significant contributions to the software itself - the only commits I see attributed are the paper and a minor tweak to deployment configurations

**Joss submission guidelines require that the submitting author is a major contributor of the software: https://joss.readthedocs.io/en/latest/submitting.html
You are correct that I have not contributed directly to the source code. However, there are many ways to contribute to software development beyond contributing to source code. On any software team, there are typically testers, documenters, project managers, source code developers, and many other roles. I have contributed over 100 hours to this project by contributing to documentation, testing, conceiving of the project, and planning the project.  I also lead the writing of the manuscript. Additionally, the corresponding author @shollingsworth, is the primary contributor to the source code.**

I'm on the fence about whether this qualifies under the JOSS substantivity requirement. On one hand, I'm definitely on board with the value of having an intuitive interface to do this work. On the other hand, I'm not fully sold on the value as currently implemented: The current service lacks substantial "workflow" tools (eg, some sort of projects/tagging, multiple collaborators, additional analytics, ...) and some level of polish/UX work to the degree that it almost feels like a Google Colab notebook could serve the purpose more efficiently (point it to a folder of images, run the nodebook, get some summary stats back out).

**I disagree with this comment. Most researchers in my field would not know how to use a Google Colab notebook to do this work. The fact that computer vision is almost solely run using programming languages is preventing its adoption in trash research, this is the premise of our work. I also am not aware of any needs they would have for these additional workflow tools that you are mentioning but would be happy to consider them if you can provide some specifics.

During the prereview some of these questions came up from @arfon as well and I had additional responses there: https://github.com/openjournals/joss-reviews/issues/5005

Please refer to the JOSS Submission guidelines here for the substantiative requirement and consider our in-line responses to why we should be considered meeting it: https://joss.readthedocs.io/en/latest/submitting.html. The guidelines state these requirements:
Age of software (is this a well-established software project) / length of commit history.
- The projects has been developed since Sep 8, 2021, approximately a year and a half
Number of commits.
- The project consists of 91 squash and merge commits which are bundled commits averaging typically around 10 commits in one, so many hundreds of commits.
Number of authors.
- The manuscript has 9 authors and there are 8 contributors to the repo demonstrating that this project has a lively open source community backing it.
Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
- The project has 57502 lines of code
Whether the software has already been cited in academic papers.
- While the project hasn't been cited in academic papers yet, we have cited several papers in the statement of need which substantiate the need in the field for this software: https://github.com/code4sac/trash-ai/blob/production/paper.md#statement-of-need
Whether the software is sufficiently useful that it is likely to be cited by your peer group.
- Same response as previous**

It also seems like the cloud deployment approach here doesn't provide much value for someone wanting to deploy it on their own - this seems like a great candidate to deploy to something like GitHub pages, Netlify, etc whereas the current approach is pretty complex for what it is. The only thing that prevents it from being deployed in that way is the collection of images to feed back to the model, which hasn't actually been followed through on yet. Even with the current structure, I'll also note that the deployment scripts and such could probably be simplified and consolidated a bit - when taking a brief look at it, it felt like I had to jump around a bit to find all the different bits and pieces and had a lot of different calls between different tools (eg, why use makefiles vs just having a single docker-compose in the root and use docker-compose up?).

**Thanks for this comment, I just created an issue for the second part of this comment: https://github.com/code4sac/trash-ai/issues/106. For the first part, I do not believe that deployment to proprietary infrastructures is within the primary scope of JOSS: https://joss.readthedocs.io/en/latest/submitting.html#submissions-using-proprietary-languages-development-environments. We made a local deployment option: https://github.com/code4sac/trash-ai/blob/production/docs/localdev.md which uses a simple docker-compose deployment. Have you tried to use this?**

Additionally, while I know the model isn't yours, when I sampled a couple images, it seems like the model is sufficiently unreliable at least to me that I'm not confident I would trust it to do the analysis for me as a researcher. I also have concerns about whether it would be fast enough just running locally to analyze the amount of images that would be needed in a research context (though not being a subject matter expert, I can't make that judgement with any real confidence).

**We did create the model and are working on it to improve it. However, the model itself isn't within the scope of a JOSS review to my knowledge, it would need to be reviewed by a machine learning journal. We are aware that it can still be improved in accuracy and that was mentioned in the video. One of the main challenges with improving its accuracy is a lack of labeled images from the diversity of possible settings that trash can be in. We are hoping that people will use the platform and share images to it so that we can relabel them and improve the model in the long term. Have you attempted to use the local deployment option? It is extremely fast and runs asynchronously so it can be run in the background. **

To be clear I definitely think this has the promise to be really useful, but I'm not sure it hits the mark as-is.

**Thanks for the kind words.**

While the video walkthrough is great, it would probably be best to have a bit more written instruction on usage in the README/on the website. I think the setup documentation could also be improved a bit. The local setup seems to refer to specifics of WSL a lot where it could be more system-agnostic, and while the usage of GH workflows is novel, the AWS deployment instructions are a bit nonstandard in that regard plus seem to assume that you're deploying it from the repo itself/as a repo owner.

**Thanks for the kind words, I created this issue to correct these issues: https://github.com/code4sac/trash-ai/issues/107 **

There doesn't seem to be any automated tests. While the desired behavior here is relatively clear/trivial, there's no formal instructions to verify behavior.

**Created this issue: https://github.com/code4sac/trash-ai/issues/108, there are many automated tests baked into the docker workflow but you are correct that they should be more clear in the documentation and there should be a formal validation proceedure.**

As tagged by editorialbot, it looks like there are some citation formatting issues

**Thanks for pointing that out. Created this issue to resolve them: https://github.com/code4sac/trash-ai/issues/109**
