### 2023-03-02 0bd7334
Cosmetic adjustments to image view screen
