### 2023-03-02 ef90b69
Removed uneeded files, added troubleshooting section
