+++ frontend/src/lib/store/canvasview.ts
@@ -1,0 +1,265 @@
+import { defineStore } from 'pinia'
+import { Rect } from '@/lib/draw'
+import type { _GettersTree } from 'pinia'
+
+export enum DialogsEnum {
+    none = 'none',
+    action = 'action',
+    label = 'label',
+}
+
+enum MouseHeldState {
+    none,
+    active,
+}
+
+enum MouseMovingState {
+    none,
+    active,
+}
+
+enum MouseHoverState {
+    none,
+    active,
+}
+
+enum MouseButton {
+    left,
+    middle,
+    right,
+    none = -1,
+}
+
+enum ObjectActionState {
+    none = 'none',
+    drawing = 'drawing',
+    resizing = 'resizing',
+    moving = 'moving',
+}
+
+enum Mode {
+    none = 'none',
+    draw = 'draw',
+    select = 'select',
+}
+
+enum Triggers {
+    none = 'none',
+    save = 'save',
+    select = 'select',
+    savenew = 'savenew',
+}
+
+interface MousePosition {
+    x: number
+    y: number
+}
+
+interface ObjectState {
+    action: ObjectActionState
+}
+
+interface MouseState {
+    held: MouseHeldState
+    position: MousePosition
+    moving: MouseMovingState
+    pressing: MouseButton
+    double_click: boolean
+    hover: MouseHoverState
+    away: boolean
+}
+
+interface ICanvas {
+    action: ObjectActionState
+    mode: Mode
+    mouse: MouseState
+    active_dialog: {
+        rect: Rect | null
+        type: DialogsEnum
+    }
+    object_status: ObjectState
+}
+
+export class CanvasMode {
+    static get Mode() {
+        return Mode
+    }
+
+    static get ObjectActionState() {
+        return ObjectActionState
+    }
+    static get MouseHeldState() {
+        return MouseHeldState
+    }
+    static get MouseMovingState() {
+        return MouseMovingState
+    }
+    static get MouseHoverState() {
+        return MouseHoverState
+    }
+    static get Triggers() {
+        return Triggers
+    }
+
+    static get MouseButton() {
+        return MouseButton
+    }
+}
+
+interface Actions {
+    set_mouse_held(boolean: boolean): void
+    set_mouse_double_click(boolean: boolean): void
+    set_action(payload: ObjectActionState): void
+    set_mouse_hover(boolean: boolean): void
+    set_dialog_active(type: DialogsEnum, rect: Rect): void
+    set_mouse_moving(boolean: boolean): void
+    set_object_status(payload: ObjectState): void
+    set_mode(payload: Mode): void
+    set_mouse_pressing(payload: MouseButton | number): void
+    conditions(): {
+        mode: { select: boolean; draw: boolean }
+        mouse: {
+            hover: boolean
+            held: boolean
+            position: MousePosition
+            moving: boolean
+            double_click: boolean
+            away: boolean
+        }
+        dialog: { action: boolean; label: boolean; none: boolean }
+        pressing: {
+            middle: boolean
+            left: boolean
+            right: boolean
+            none: boolean
+        }
+        action: {
+            drawing: boolean
+            resizing: boolean
+            none: boolean
+            moving: boolean
+        }
+    }
+    set_mouse_position(payload: MousePosition): void
+    set_mouse_away(boolean: boolean): void
+}
+
+interface Getters extends _GettersTree<ObjectState> {}
+
+export const useCanvas = defineStore<'canstore', ICanvas, Getters, Actions>(
+    'canstore',
+    {
+        state: () =>
+            ({
+                action: ObjectActionState.none,
+                mode: Mode.none,
+                active_dialog: {
+                    rect: null,
+                    type: DialogsEnum.none,
+                },
+                mouse: {
[…]
