### 2023-03-06 code4sac/trash-ai#122 @wincowgerDEV
[Feature]: Need to be more explicit about data sharing
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Description

Reviewer said: While it is stated in the paper that some of the images are uploaded to an S3 storage there is no information about it on your trashai.org webpage. I think it is at least necessary to put a disclaimer to users that their image data is uploaded and that images may contain delicate information through exif data (because I think the targeted audience is not always aware of what information images may contain). In my view it would even be better to have an opt-out option for image upload and/or stripping unnecessary parts of the exif data while uploading and/or informing users of uploading process to aws while also putting information resources to exif data directly on the page.

### Problem

Not clear enough information about what happens to data when it goes to the website. Users don't have a way to opt out of data sharing.

### Proposed Solution

Add popup or info box near the data upload that tells the user where their data will be shared and allows them or opt out.

### Alternatives Considered

Expect users to read the documentation.
