### 2023-03-06 code4sac/trash-ai#123 @wincowgerDEV
[Feature]: Add example dataset for users to upload to try to tool and to inspect the json output
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Description

Reviewer Said: I find your video tutorial very nice and good to follow. Additionally, I think it would be nice if you'd also have a folder of example data with which new users could directly follow your tutorial. Maybe even with a post-processing example of json data.

### Problem

Users may not have access to large datasets of trash images or may want to debug or inspect the json before using the tool.

### Proposed Solution

Create an OSF database full of images that a users could upload to the tool to test it and download cache the output of the tool so that users can compare and inspect the output.

### Alternatives Considered

Expect users have data already and will use the tool without inspection.
