### 2023-03-06 code4sac/trash-ai#124 @wincowgerDEV
[Feature]: Data Structure Update
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Description

Reviewer Said:
I agree with @luxaritas that the written documentation could be expanded. What I miss the most is a detailed explanation of the json data structure for post-processing software. Which fields are expected and what information do they contain? Are there fields which are not available for all data. As I am not a familiar with trash research I was also wondering whether there is some standardized format for data exchange on trash location data and labelling which could be used here (I think the Hapich et al 2022 paper is elaborating on this, so it would be nice to read some more details of the connection of this trash taxonomy to the data format of trashai). Also I understand that your targeted audience is less technical so dealing with json data may be a barrier. I think it would also be good to offer a pdf based overview (like an analysis page which could just be printed into a pdf - so users could directly have a map overview of their trash expedition).

### Problem

JSON Schema isn't well defined. JSON format can be challenging to deal with and summary stats may be preferred.

### Proposed Solution

Create a json schema file and some simple to read documentation on the json format. Also create a download with summary information in csv of similar flat format.

### Alternatives Considered

Expect that folks are savy in json.
