### 2023-03-06 #5136 @wincowgerDEV
@domna Thanks so much for the comments and thorough review. Some responses in line below.

> I agree with @luxaritas view on the deployment, that the software should be easily deployable to any hosting solution or local deployment. While you provide a localdev option I see it, however, as an option for doing local development for an aws based system instead of as a full-fledged deployment solution. From my view the localstack containers and backend upload options are not really necessary for a local deployment.
What I would find nice to have would be:

> Documentation on how to compile the front-end and directly upload it to any hosting solution (ideally with disabled backend upload option, so you really only need a webspace or even could use gh pages).
- Could you provide an example of an application that has this kind of functionality? I am not super familiar with it but definitely interested in figuring out how we can do it.

> A single docker container which contains the compiled gui so admins can easily deploy in a docker based system. It would also be nice if this container would be uploaded to gh packages or dockerhub, so you could just provide a docker-compose file people could use to deploy the service.
- Is this something different than what we already have for docker-compose deployment? https://github.com/code4sac/trash-ai/blob/production/docs/localdev.md#option-2-using-the-shell

> Privacy

> While it is stated in the paper that some of the images are uploaded to an S3 storage there is no information about it on your trashai.org webpage. I think it is at least necessary to put a disclaimer to users that their image data is uploaded and that images may contain delicate information through exif data (because I think the targeted audience is not always aware of what information images may contain). In my view it would even be better to have an opt-out option for image upload and/or stripping unnecessary parts of the exif data while uploading and/or informing users of uploading process to aws while also putting information resources to exif data directly on the page.
- Agree with you 100% on this. Just made an issue to fix it. https://github.com/code4sac/trash-ai/issues/122

> Example data / Tutorial

> I find your video tutorial very nice and good to follow. Additionally, I think it would be nice if you'd also have a folder of example data with which new users could directly follow your tutorial. Maybe even with a post-processing example of json data.
- Thanks for the kind words. Definitely agree here too and have made this issue to deal with it. https://github.com/code4sac/trash-ai/issues/123

> Documentation and data structure

> I agree with @luxaritas that the written documentation could be expanded. What I miss the most is a detailed explanation of the json data structure for post-processing software. Which fields are expected and what information do they contain? Are there fields which are not available for all data. As I am not a familiar with trash research I was also wondering whether there is some standardized format for data exchange on trash location data and labelling which could be used here (I think the Hapich et al 2022 paper is elaborating on this, so it would be nice to read some more details of the connection of this trash taxonomy to the data format of trashai). Also I understand that your targeted audience is less technical so dealing with json data may be a barrier. I think it would also be good to offer a pdf based overview (like an analysis page which could just be printed into a pdf - so users could directly have a map overview of their trash expedition).
- Thanks for this recommendation. It is extremely helpful and I have made an issue for it here: https://github.com/code4sac/trash-ai/issues/124

> Reference to other software

> I was missing information on which software people typically use for researching trash images. Since I'm not familiar with the field I may not be aware that such software is not really available, but I think it would be useful to give a glimpse into the working process around your tool (like: what do I actually do with the downloaded json data?). It would also be interesting to know whether there are some widely known databases where you can put your trash data, as I think the useful part of such information is putting a lot of data from different researchers together to give a broad overview of littering places. I think especially it would be nice to have some guidance what to do with the json data downloaded by your tool. Is there some analysis software for that?
- Great idea! Added in these here: https://github.com/code4sac/trash-ai/issues/124

> I was also very happy to read that you plan to upload your data to the taco dataset. Do you also plan on adding a feature where users can directly annotate their data in trashai and then upload the annotation together with the images to trashai. I think this could be a powerful tool of advancing the taco dataset, since it would target the erroneous classifications of the model directly.
- Thanks for the kind words. We do plan on adding in that feature and are actively working on it. Glad to hear you also think it will be useful.

> Future of the software

> This is just more out of curiosity for your future plans. I find your software a super helpful tool for annotation. However, I was wondering if you plan to bring the data of different sources together. As you state in your paper such information is often used for policy decisions and I think the approach is really useful when data from different places come together. So going in the direction of having a database for trash researchers where data can be easily uploaded from trashai and policy makers can search for locations to get information in this area could be extremely helpful. Are you already thinking in this direction?
- Love this vision you are sharing. We don't currently have a central repository for trash images and historically image data hasn't been shared very much in our field even though they have been used in many studies. I will think on this some more to consider how best to integrate the tools. We have a portal we have been developing for sharing images and other data in a standardized format of microplastics wincowger.shinyapps.io/validator. It could be used for trash images and data as well but we haven't extended the schema for it yet. Will think about this for the long term roadmap of the tool for sure.
