### 2023-03-06 code4sac/trash-ai#125 @wincowgerDEV
[Feature]: Documentation on the ML model
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Description

Reviewer said: To be clear, I'm not _necessarily_ advocating for this as an addition to the GUI - the reason I brought it up was more around trying to figure out the value proposition of the UI itself and whether it was providing something that couldn't be done in a simpler way, and it sounds like the UI is important to make CV tools accessible to researchers. Integration with programable interfaces could still be valuable though (at the very least, I'd imagine it would be a good idea for the model to be available directly and not _just_ via UI).

### Problem

Some people likely want to use the model without the tool.

### Proposed Solution

Add documentation about how to use the model without the tool or at least where it is.

### Alternatives Considered

Assume people find the ML without documentation.
