### 2023-03-06 #5136 @wincowgerDEV
@luxaritas Thanks for the follow up clarification.

> To be clear, I'm not _necessarily_ advocating for this as an addition to the GUI - the reason I brought it up was more around trying to figure out the value proposition of the UI itself and whether it was providing something that couldn't be done in a simpler way, and it sounds like the UI is important to make CV tools accessible to researchers. Integration with programable interfaces could still be valuable though (at the very least, I'd imagine it would be a good idea for the model to be available directly and not _just_ via UI).
- I definitely agree with you that some people will prefer something like Colab. Also super good idea to add documentation about where the model is and how people can use it without the tool. Added an issue for it. https://github.com/code4sac/trash-ai/issues/125

> The confidence labels are definitely a great idea. At some level though model accuracy itself is a barrier to usefulness of this tool - no matter what else you do, if you have to manually re-tag all the images anyways it's not helping automate the process which is the whole point... Unfortunately it's a weird situation where the model itself isn't the primary thing under scrutiny for JOSS, but it's a critical part of the application's functionality
- I agree with you that accuracy is a barrier to usability. It is going to take us a long time to build a general trash ai but I don't think that precludes us from getting this publication out and allowing people to begin using this tool. It is actually already being used by state agencies around California and a country wide trash survey: https://www.notracetrails.com/trashteam. In the end, it is a double edged sword. We need people to use the tool so that we can get the data we need to improve it but they are less likely to use it until it performs better. In the long term, we expect it to be a standard in trash research.
