### 2023-03-07 code4sac/trash-ai#126 @wincowgerDEV
[Feature]: Static JS bundle support
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Description

I just meant that it would be good to have a documentation on how to generate a static javascript bundle and upload it to a self-hosted webspace (e.g. locally for people wanting to use this model in an intranet or so).
Specifically, this means having something like

cd frontend
yarn build
cp dist <your-webspace-folder>

in the documentation. Regarding the disabled backend I've seen that it is possible to just use the frontend without any issues while testing this.
For gh pages it would just mean that you have an automatic gh action to generate a webpage which would then compile the js and copy it. You can simply use this action here https://github.com/marketplace/actions/github-pages-action and add your yarn build step. I think it should just work. The benefit would mainly be for development purposes, so people cloning the repo could immediately rollout the changes through this action.

### Problem

Slow to deploy from docker and may not be necessary for all use cases.

### Proposed Solution

cd frontend
yarn build
cp dist <your-webspace-folder>

### Alternatives Considered

Expect Docker Deployment.
