### 2023-04-06 code4sac/trash-ai#127 @shollingsworth
@feydan Does the org have a docker registry?
