### 2023-04-12 baa4616
Create generic.md
I'm fixing the generic issue template I submitted before. This one should work!
