### 2023-04-13 0ca0da0
Delete generic.md (#139)
