+++ .github/ISSUE_TEMPLATE/generic.md
@@ -1,13 +1,0 @@
-### Dependencies
-ANY ISSUE NUMBERS THAT ARE BLOCKERS OR OTHER REASONS WHY THIS WOULD LIVE IN THE ICEBOX - DELETE THIS SECTION IF NO DEPENDENCIES
-
-### Overview
-WE NEED TO DO X FOR Y REASON
-
-### Action Items
-A STEP BY STEP LIST OF ALL THE TASK ITEMS THAT YOU CAN THINK OF NOW EXAMPLES INCLUDE: Research, reporting, etc.
-
-- [ ] INSERT ITEMS ...
-
-### Resources/Instructions
-REPLACE THIS TEXT -If there is a website which has documentation that helps with this issue provide the link(s) here.
