### 2023-04-13 7724192
Shollingsworth patch 1 (#142)
* Update generic_issue.yml

* Update generic_issue.yml
