+++ .github/ISSUE_TEMPLATE/generic_issue.yml
@@ -1,7 +1,6 @@
 name: Generic Issue
 description: Organized way of making actionable tasks for PMs, data scientists, leads, etc.
 title: "<title>"
-labels:
 body:
   - type: checkboxes
     id: Terms
@@ -42,6 +41,6 @@ body:
       label: Anything else?
       description: |
         Links? References? Anything that will give us more context about the task and how to accomplish it.
-      Tip: You can attach images or log files by clicking this area to highlight it and then dragging files in.
-      validations:
-        required: false
+      placeholder: You can attach images or log files by clicking this area to highlight it and then dragging files in.
+    validations:
+      required: false
