+++ .github/ISSUE_TEMPLATE/generic_issue.yml
@@ -1,6 +1,6 @@
 name: Generic Issue
 description: Organized way of making actionable tasks for PMs, data scientists, leads, etc.
-title: "<title>"
+title: "[Generic]: "
 body:
   - type: checkboxes
     id: Terms
