### 2023-04-20 code4sac/trash-ai#123 @feydan
![image](https://user-images.githubusercontent.com/4978032/233241611-d679e946-8966-4f7b-bbfc-5cc376b6ad93.png)
I think this is already done @wincowgerDEV .  When you haven't uploaded any images, there's a "Show Samples" button.
