### 2023-04-20 87f3849
Adding publication to projection information
