### 2023-05-04 ef38e4c
Updated Contributor section of the readme (#154)
* Coco POC #8 #9 (#19) (#26)

* Fixed some package.json issues, updated some entries in .gitignore and .in (for those who would use it)

* saving work

* saving work... having issues with async reading file then loading it in coco

* coco confirmed working

* Added JSON metadata

* Arg! this should work

* Removed Pro port mapping from localstack docker-compose

* Added multiple file upload and cleaned up presentation a bit

Looks like there is a problem with uploading multiple files at once, the
predictions don't always trigger

* Minor updates
[…]
