### 2023-05-10 86e09ef
🎉 feat(frontend): add Dockerfile and docker-compose.yml for frontend
