+++ frontend/docker-compose.yml
@@ -1,0 +1,6 @@
+version: '3'
+services:
+  frontend:
+    build: .
+    ports:
+      - "8080:80"
