### 2023-05-10 4a86f61
Refactor frontend Dockerfile
