+++ frontend/Dockerfile
@@ -30,5 +30,3 @@ EXPOSE 80
 # Start Nginx
 CMD ["nginx", "-g", "daemon off;"]
 
-# This Dockerfile starts with the official Node.js 14 Alpine image as the base image. It sets the working directory to `/app`, copies the `package.json` and `yarn.lock` files to the container, and installs the dependencies using Yarn.
-# Next, it copies the rest of the application code to the container and builds the application using Vite. It exposes port 3000 for the application and starts the application using the production build.
