### 2023-05-11 12a65eb
127 feature docker compose only deployment (#157)
* Add tags to ignore. (commit message: "Ignore tags")

* 🎉 feat(frontend): add Dockerfile and docker-compose.yml for frontend

* Refactor frontend Dockerfile
