### 2023-05-11 code4sac/trash-ai#127 @shollingsworth
There are two remaining items on this:
1.)  pushing to docker registry during frontend build
2.) Fix the google api key during build process
3.) update README to reflect changes, remove uneeded localdev pieces because backend isn't being called right now.
