+++ .github/workflows/dockerhub.yaml
@@ -1,0 +1,36 @@
+name: Dockerhub Publish
+
+on:
+  workflow_dispatch:
+
+permissions:
+  id-token: write
+  contents: read
+
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Git clone the repository
+        uses: actions/checkout@v1
+
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v2
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v2
+
+      - name: Login to Docker Hub
+        uses: docker/login-action@v2
+        with:
+          username: code4sac
+          password: ${{ secrets.CODE4SAC_DOCKER_API_KEY }}
+
+      - name: Build and push
+        uses: docker/build-push-action@v4
+        env:
+          VITE_GOOGLE_MAPS_API_KEY: ${{ secrets.VITE_GOOGLE_MAPS_API_KEY }}
+        with:
+          context: frontend
+          push: true
+          tags: code4sac/trashai:latest
