+++ frontend/Dockerfile
@@ -29,4 +29,3 @@ EXPOSE 80
 
 # Start Nginx
 CMD ["nginx", "-g", "daemon off;"]
-
