### 2023-05-11 381dbfd
Adjusted nginx to handle SPA app, testing secret configuration
