+++ .github/workflows/dockerhub.yaml
@@ -8,7 +8,6 @@ on:
     paths:
       - "frontend/**"
 
-
 permissions:
   id-token: write
   contents: read
