### 2023-05-12 29c14cf
Fixes #127 - removed unnecessary backend components, added docker run to README
