### 2023-05-12 code4sac/trash-ai#127 @shollingsworth
All fixes have been implemented. Closing issue.
