### 2023-05-12 cf6c4a3
Fixes #127 - removed unnecessary backend components, added docker run to README (#159)
