### 2023-05-12 c412a24
Addresses #107 - Added other web server deployment to README based on public docker image
