### 2023-05-12 5f7f017
Fixes #127 #107 - deployment and run docker instances (#160)
* Fixes #127 - removed unnecessary backend components, added docker run to README

* Addresses #107 - Added other web server deployment to README based on public docker image
