### 2023-05-12 code4sac/trash-ai#107 @shollingsworth
See attached pull Request https://github.com/code4sac/trash-ai/pull/160
