### 2023-05-12 code4sac/trash-ai#106 @shollingsworth
Minus the aws/github stuff can this be closed with #127 and #107 ?
