### 2023-05-12 code4sac/trash-ai#126 @shollingsworth
Minus the aws/github stuff can this be closed with https://github.com/code4sac/trash-ai/issues/127 and https://github.com/code4sac/trash-ai/issues/107 ?
