### 2023-05-12 code4sac/trash-ai#122 @shollingsworth
We might want to update the current paper to say we're not actually doing any interaction with the backend as of now. In its current state it's only a web frontend. In the future we may implement this, but it's not a factor at the moment.
