### 2023-05-12 code4sac/trash-ai#124 @wincowgerDEV
> I'm not sure what this is asking, might need to get with Win on this one.

@shollingsworth, yeah this one is in my court. I can do all of this.
