### 2023-05-12 2527ec1
Update paper.md
https://github.com/code4sac/trash-ai/issues/122
