### 2023-05-12 code4sac/trash-ai#122 @wincowgerDEV
Updated the paper 2527ec10682118e6ba60a4dba2026e0460e82b73, now put this as a future task since it wasn't implemented.
