### 2023-05-16 code4sac/trash-ai#124 @wincowgerDEV
Here is the json schema for the summary file. I think it should be added in the download.

`{
  "$schema": "http://json-schema.org/draft-07/schema#",
  "type": "object",
  "properties": {
    "detected_objects": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "count": { "type": "integer", "minimum": 0 },
          "hashes": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["name", "count", "hashes"]
      }
    },
    "no_detection_hashes": {
      "type": "array",
      "items": { "type": "string" }
    },
    "unique_detections": { "type": "integer", "minimum": 0 },
    "total_detections": { "type": "integer", "minimum": 0 },
    "gps": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "coordinate": {
                "type": "object",
                "properties": {
                  "lat": { "type": "number" },
                  "lng": { "type": "number" }
                },
                "required": ["lat", "lng"]
              },
              "hash": { "type": "string" }
            },
            "required": ["coordinate", "hash"]
          }
        }
      },
      "required": ["list"]
    }
  },
  "required": ["detected_objects", "no_detection_hashes", "unique_detections", "total_detections", "gps"]
}
`

Here is a description file:

1. `detected_objects` (Array): An array of objects where each object represents a detected item with the following properties:
   - `name` (String): The name of the detected item (e.g., "Plastic film").
   - `count` (Integer): The count of the detected item. This number should be a non-negative integer.
   - `hashes` (Array): An array of strings where each string represents a hash associated with the detected item.

2. `no_detection_hashes` (Array): An array of strings where each string represents a hash that is associated with items not detected.

3. `unique_detections` (Integer): The count of unique detections. This number should be a non-negative integer.

4. `total_detections` (Integer): The total count of detections. This number should be a non-negative integer.

5. `gps` (Object): An object that contains GPS information with the following properties:
   - `list` (Array): An array of objects where each object represents a GPS location with the following properties:
     - `coordinate` (Object): An object that contains the GPS coordinates with the following properties:
       - `lat` (Number): The latitude of the location.
       - `lng` (Number): The longitude of the location.
     - `hash` (String): A hash associated with the location.

Please note that all of the properties described above are required. If any property is missing or if a property's value is of an incorrect type, the JSON will not validate against the schema.
