### 2023-05-16 code4sac/trash-ai#124 @wincowgerDEV
> Adding analysis page:
already exists https://www.trashai.org/summary
