### 2023-05-16 code4sac/trash-ai#124 @wincowgerDEV
Todo:
- [x] Add JSON schema and definition files to the download
- [x] Add an option to download data as csv instead of json.
- [ ] Add video discussion about things a user can do with the data.
