### 2023-05-16 code4sac/trash-ai#123 @wincowgerDEV
Yup, good catch, closing this issue.
