### 2023-05-16 code4sac/trash-ai#125 @wincowgerDEV
I think this should be part of a much longer-term vision for the tool. Closing for now and will let reviewers know this isn't intended at this time.
