### 2023-05-16 code4sac/trash-ai#106 @wincowgerDEV
You're a legend @shollingsworth! Yeah, I think this is settled, closing now.
