### 2023-05-16 code4sac/trash-ai#126 @wincowgerDEV
I think we have made a lot of progress on this and will let the reviewers know we are where we would like to be for deployment ease.
