### 2023-05-16 code4sac/trash-ai#108 @wincowgerDEV
@shollingsworth is it trivial to write some instructions to verify the automated tests? Another thing I could probably do instead is just make a test set and then make a video explaining how someone could use it to verify behavior.
