### 2023-05-16 code4sac/trash-ai#108 @wincowgerDEV
Perhaps adding this is actually enough: https://github.com/code4sac/trash-ai/issues/124

If we add that to the downloads and then I explain how to use the schema validator tool like https://www.jsonschemavalidator.net/ in the updated video then that should be proof enough that the pipeline is working as expected to create a data output useful for users.
