### 2023-06-01 #5136 @wincowgerDEV
Hey @luxaritas,
I want to make sure that we can meet your expectations for automated tests:
> There doesn't seem to be any automated tests. While the desired behavior here is relatively clear/trivial, there's no formal instructions to verify behavior.

We are thinking about doing the following:
1) Behavior of the app stability is currently being verified by github actions which would throw an error automatically if something were wrong with the deployment.
2) We will write out instructions for using json schemas to validate the data output format and provide a tutorial on feeding a test example to the tool and receiving an expected output that the user can test against.

Question:
Do you think these tests are enough or would you like to see additional specific automated tests?
