### 2023-06-01 #5136 @luxaritas
The JOSS requirement is:
>  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

The way I personally interpret that is: preferably with an automated script, but worst case with specific instructions for testing "by hand", you should be able to ensure the key portions of the app's functionality behave as expected. Your current GitHub actions ensures that the deployment is successful, but as far as I can tell does not validate anything about the actual functionality of the application. Things that would be good to verify could include:
* Does the landing page load?
* On hitting show samples, does processing complete in an expected amount of time? Does the result page show all results expected?
* On hitting upload and providing images, does processing complete in an expected amount of time? Does the result page show all results expected?
* After receiving results, can you click an image and does the page show the expected data/metadata/etc?
* After receiving results, does the summary page contain the correct data?

When I went to look at the application right now, it actually looked like it wasn't processing any data when running the samples - with formal tests, it should ensure you catch that before changes are launched!

If you wanted to automate that, you might consider spinning up a local environment in GitHub actions and then using something like Cypress/Selenium/etc to interact with the page and verify it behaves as you expect, though per the JOSS requirements minimally there should be well-defined test criteria. How you perform the tests and how many specific test cases you have and their granularity (just end-to-end tests? unit tests? both?) is ultimately up to you, but the idea is for you to have higher confidence in making changes, catch issues quicker, and make it easier for new contributors to feel comfortable making changes - automation helps amplify those benefits because you can just run your validation with a single command, and even do such "smoke checks" on every commit/PR/etc without having to manually perform those checks.
