### 2023-06-01 #5136 @wincowgerDEV
Thanks for this follow-up.  We greatly appreciate this detailed feedback. It is extremely helpful for us moving forward to think about how best to automate the testing of our app. I will take these recommendations to the web guru and see what he thinks he can implement before we resubmit this publication. Agreed that having a separate service that tests the client-side experience would be ideal for ensuring performance but I also know these can be exceptionally challenging to implement and maintain so I don't want to overburden the devs, we are all volunteers.
