### 2023-06-01 code4sac/trash-ai#108 @wincowgerDEV
Got some additional feedback from the reviewers on this: https://github.com/openjournals/joss-reviews/issues/5136#issuecomment-1571293951

@shollingsworth I think they are ok with us doing manual tests, but they would prefer we implement some automated user-side tests on the website/container using selenium or something similar if it isn't too crazy to implement.
