### 2023-06-05 716d0af
Update README.md
Added instruction for running locally on an Apple Silicon device.
