### 2023-06-07 8c56b8f
[Feature]: Document model/build tools (#166)
* Add basic documentation

* Add Requirements

* Add clarification
