### 2023-06-07 code4sac/trash-ai#108 @Zesky665
@wincowgerDEV I read the suggestion in the JOSS thread. I can implement all of those tests using something like Cypress, it shouldn't be much of a hassle.
