### 2023-06-07 code4sac/trash-ai#108 @wincowgerDEV
That would be amazing! Thanks so much.
