+++ notebooks/README.md
@@ -1,0 +1,3 @@
+# Other work
+
+- [Running-man-01/trashai_nbs](https://github.com/running-man-01/trashai_nbs) is a sister project for TrashAI that explores different models with this project
