### 2023-06-14 43701ba
Add smoke test draft
