+++ frontend/jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    globalSetup: './setup.js',
+    globalTeardown: './teardown.js',
+    testEnvironment: './puppeteer_environment.js',
+}
