### 2023-06-16 d827480
Added back manual instructions
