### 2023-06-16 064d300
Add instruction for running auto tests
