### 2023-06-22 4ab321a
adding schema files to the zip download
