### 2023-06-26 7c96b91
Adding schema files to download (#173)
* adding schema files to the zip download

* revert yarn change

* fixing list

* indentation
