+++ frontend/src/lib/util.ts
@@ -137,3 +137,5 @@ export const getStaticFiles = (filePaths: string[]) =>
                 .then((blob) => ({ filePath: path, blob })),
         ),
     )
+
+
