+++ paper.bib
@@ -88,8 +88,7 @@
   number   =  1,
   pages    = "15",
   month    =  jun,
-  year     =  2022,
-  doi      =  10.1186/s43591-022-00035-1
+  year     =  2022
 }
 
 @misc{Waterboards:2018,
@@ -109,7 +108,7 @@
     number = {8},
     pages = {e2019EA000960},
     keywords = {plastic pollution, object detection, automated monitoring, deep learning, artificial intelligence, river plastic},
-    doi = {10.1029/2019EA000960},
+    doi = {https://doi.org/10.1029/2019EA000960},
     url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2019EA000960},
     eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2019EA000960},
     note = {e2019EA000960 10.1029/2019EA000960},
@@ -145,8 +144,7 @@
   number   =  1,
   pages    = "6",
   month    =  jun,
-  year     =  2018,
-  doi      =  10.1186/s40965-018-0050-y
+  year     =  2018
 }
 
 
@@ -177,10 +175,11 @@ volume = {138},
 pages = {274-284},
 year = {2022},
 issn = {0956-053X},
-doi = {10.1016/j.wasman.2021.12.001},
+doi = {https://doi.org/10.1016/j.wasman.2021.12.001},
 url = {https://www.sciencedirect.com/science/article/pii/S0956053X21006474},
 author = {Sylwia Majchrowska and Agnieszka Mikołajczyk and Maria Ferlin and Zuzanna Klawikowska and Marta A. Plantykow and Arkadiusz Kwasigroch and Karol Majek},
-keywords = {Object detection, Semi-supervised learning, Waste classification benchmarks, Waste detection benchmarks, Waste localization, Waste recognition}
+keywords = {Object detection, Semi-supervised learning, Waste classification benchmarks, Waste detection benchmarks, Waste localization, Waste recognition},
+abstract = {Waste pollution is one of the most significant environmental issues in the modern world. The importance of recycling is well known, both for economic and ecological reasons, and the industry demands high efficiency. Current studies towards automatic waste detection are hardly comparable due to the lack of benchmarks and widely accepted standards regarding the used metrics and data. Those problems are addressed in this article by providing a critical analysis of over ten existing waste datasets and a brief but constructive review of the existing Deep Learning-based waste detection approaches. This article collects and summarizes previous studies and provides the results of authors’ experiments on the presented datasets, all intended to create a first replicable baseline for litter detection. Moreover, new benchmark datasets detect-waste and classify-waste are proposed that are merged collections from the above-mentioned open-source datasets with unified annotations covering all possible waste categories: bio, glass, metal and plastic, non-recyclable, other, paper, and unknown. Finally, a two-stage detector for litter localization and classification is presented. EfficientDet-D2 is used to localize litter, and EfficientNet-B2 to classify the detected waste into seven categories. The classifier is trained in a semi-supervised fashion making the use of unlabeled images. The proposed approach achieves up to 70% of average precision in waste detection and around 75% of classification accuracy on the test dataset. The code and annotations used in the studies are publicly available online11https://github.com/wimlds-trojmiasto/detect-waste..}
 }
 
 @misc{Proença:2020,
@@ -193,4 +192,3 @@ keywords = {Object detection, Semi-supervised learning, Waste classification ben
   year = {2020},
   copyright = {arXiv.org perpetual, non-exclusive license}
 }
-
