+++ paper.bib
@@ -88,7 +88,8 @@
   number   =  1,
   pages    = "15",
   month    =  jun,
-  year     =  2022
+  year     =  2022,
+  doi      =  "10.1186/s43591-022-00035-1"
 }
 
 @misc{Waterboards:2018,
@@ -108,7 +109,7 @@
     number = {8},
     pages = {e2019EA000960},
     keywords = {plastic pollution, object detection, automated monitoring, deep learning, artificial intelligence, river plastic},
-    doi = {https://doi.org/10.1029/2019EA000960},
+    doi = {10.1029/2019EA000960},
     url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2019EA000960},
     eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2019EA000960},
     note = {e2019EA000960 10.1029/2019EA000960},
@@ -144,7 +145,8 @@
   number   =  1,
   pages    = "6",
   month    =  jun,
-  year     =  2018
+  year     =  2018,
+  doi      =  "10.1186/s40965-018-0050-y"
 }
 
 
@@ -175,7 +177,7 @@ volume = {138},
 pages = {274-284},
 year = {2022},
 issn = {0956-053X},
-doi = {https://doi.org/10.1016/j.wasman.2021.12.001},
+doi = {10.1016/j.wasman.2021.12.001},
 url = {https://www.sciencedirect.com/science/article/pii/S0956053X21006474},
 author = {Sylwia Majchrowska and Agnieszka Mikołajczyk and Maria Ferlin and Zuzanna Klawikowska and Marta A. Plantykow and Arkadiusz Kwasigroch and Karol Majek},
 keywords = {Object detection, Semi-supervised learning, Waste classification benchmarks, Waste detection benchmarks, Waste localization, Waste recognition},
