### 2023-06-27 03b50dd
update paper acknowledgments and links.
