### 2023-06-27 code4sac/trash-ai#108 @wincowgerDEV
@Zesky665 should this be closed now thanks to your work in 8c56b8f2c94494fd4ab3f1dbc7c4031fdbc29afb?
