### 2023-06-27 code4sac/trash-ai#108 @Zesky665
@wincowgerDEV No, me and @shollingsworth still need to add the auto tests to the deployment workflow. We can declare it done after that.
