### 2023-06-28 code4sac/trash-ai#108 @wincowgerDEV
Ok, sounds good.
