### 2023-07-01 fedadc3
Simplifying generateSummaryFiles and removing array mutation
