### 2023-07-01 0fe7dab
Consistent csv delimiter (#179)
* Only using comma delimited for summary csv files and automatically quoting all string fields in the csv

* Simplifying generateSummaryFiles and removing array mutation

* fixing unit test

* only run typescript tests with test-ts

* Updated babel and dev package to run tests

---------

Co-authored-by: Steven Hollingsworth <749515+shollingsworth@users.noreply.github.com>
