+++ frontend/jest.config.js
@@ -1,5 +1,6 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-};
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    testMatch: ['**/*.spec.ts'],
+}
