+++ frontend/src/lib/csv/schema-parser.spec.ts
@@ -73,11 +73,11 @@ describe('csv', () => {
             ]),
         )(obj)
         expect(csv).toEqual(
-            'id,fruit,cultivar,inSeason\n' +
-                '1,apple,gala,false\n' +
-                '1,apple,red delicious,true\n' +
-                '1,pear,bosc,false\n' +
-                '1,pear,bartlett,true',
+            '"id","fruit","cultivar","inSeason"\n' +
+                '1,"apple","gala",false\n' +
+                '1,"apple","red delicious",true\n' +
+                '1,"pear","bosc",false\n' +
+                '1,"pear","bartlett",true',
         )
     })
 
