+++ frontend/babel.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-    presets: ['@vue/cli-plugin-babel/preset'],
+    presets: ['@vue/cli-plugin-babel/preset', '@babel/preset-typescript'],
 }
