### 2023-07-22 402e595
Review1 (#181)
* my workflow

* example templates for manuscript.

* manuscript info

* update citations

* cleanup text

* cite

* update readme

* contribute

* add website badge

* update web link

* Update README.md

* Update README.md

* add video link

* auto update

* utility files device specific

* move the pdf generator to workflows

* test error in paper.bib

* test citation style

* add citations

* add a few more citations

* try without space

* add all

* no spaces allowed in citation names

* add demo images

* Update paper.md

* give images some space

* image captions

* add 
[…]
