+++ paper.bib
@@ -88,7 +88,8 @@
   number   =  1,
   pages    = "15",
   month    =  jun,
-  year     =  2022
+  year     =  2022,
+  doi      =  "10.1186/s43591-022-00035-1"
 }
 
 @misc{Waterboards:2018,
@@ -108,11 +109,10 @@
     number = {8},
     pages = {e2019EA000960},
     keywords = {plastic pollution, object detection, automated monitoring, deep learning, artificial intelligence, river plastic},
-    doi = {https://doi.org/10.1029/2019EA000960},
+    doi = {10.1029/2019EA000960},
     url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2019EA000960},
     eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2019EA000960},
     note = {e2019EA000960 10.1029/2019EA000960},
-    abstract = {Abstract Quantifying plastic pollution on surface water is essential to understand and mitigate the impact of plastic pollution to the environment. Current monitoring methods such as visual counting are labor intensive. This limits the feasibility of scaling to long-term monitoring at multiple locations. We present an automated method for monitoring plastic pollution that overcomes this limitation. Floating macroplastics are detected from images of the water surface using deep learning. We perform an experimental evaluation of our method using images from bridge-mounted cameras at five different river locations across Jakarta, Indonesia. The four main results of the experimental evaluation are as follows. First, we realize a method that obtains a reliable estimate of plastic density (68.7\% precision). Our monitoring method successfully distinguishes plastics from environmental elements, such as water surface reflection and organic waste. Second, when trained on one location, the method generalizes well to new locations with relatively similar conditions without retraining (≈50\% average precision). Third, generalization to new locations with considerably different conditions can be boosted by retraining on only 50 objects of the new location (improving precision from ≈20\% to ≈42\%). Fourth, our method matches visual counting methods and detects ≈35\% more plastics, even more so during periods of plastic transport rates of above 10 items per meter per minute. Taken together, these results demonstrate that our method is a promising way of monitoring plastic pollution. By extending the variety of the data set the monitoring method can be readily applied at a larger scale.},
     year = {2020}
 }
 
@@ -145,7 +145,8 @@
   number   =  1,
   pages    = "6",
   month    =  jun,
-  year     =  2018
+  year     =  2018,
+  doi      =  "10.1186/s40965-018-0050-y"
 }
 
 
@@ -176,7 +177,7 @@ volume = {138},
 pages = {274-284},
 year = {2022},
 issn = {0956-053X},
-doi = {https://doi.org/10.1016/j.wasman.2021.12.001},
+doi = {10.1016/j.wasman.2021.12.001},
 url = {https://www.sciencedirect.com/science/article/pii/S0956053X21006474},
 author = {Sylwia Majchrowska and Agnieszka Mikołajczyk and Maria Ferlin and Zuzanna Klawikowska and Marta A. Plantykow and Arkadiusz Kwasigroch and Karol Majek},
 keywords = {Object detection, Semi-supervised learning, Waste classification benchmarks, Waste detection benchmarks, Waste localization, Waste recognition},
@@ -193,4 +194,3 @@ abstract = {Waste pollution is one of the most significant environmental issues 
   year = {2020},
   copyright = {arXiv.org perpetual, non-exclusive license}
 }
-
