+++ docs/localdev.md
@@ -16,11 +16,11 @@ These values can be adjusted by editing the localdev env file [.env](../localdev
 It's suggested you work in branch `local` by creating your own local branch when developing
 Pushing / merging PR's to any branches with a prefix of `aws/` will trigger deployment actions
 For full functionality you will want to get a Google Maps API key and name it VITE_GOOGLE_MAPS_API_KEY, but it is not required
-=======
 
 Pushing / merging PR's to any branches with a prefix of `aws/` will
 trigger deployment actions, when developing locally, create a new branch
 and submit a pull request to `aws/trashai-staging`
+
 
 ---
 # Set up 

[…]
