### 2023-08-08 code4sac/trash-ai#108 @wincowgerDEV
Closing this because we added manual tests now.
