### 2023-08-08 #5136 @wincowgerDEV
Wanted to give another update here. We got stalled out doing the automated tests because it was too challenging for our team to implement but have implemented quite a few manual tests and the docker compiler is handling the vast majority of the existential tests. All updates have been incorporated into the live repo https://github.com/code4sac/trash-ai. We believe that we have satisfactorily addressed all JOSS requirements and did our best to implement the additional ideas for improvement proposed by the reviewers.

You can see the issues which we resolved specifically pertaining to the reviewer's comments here: https://github.com/code4sac/trash-ai/issues?q=label%3A%22reviewer+comment%22+is%3Aclosed

The majority of the reviewers comments were resolved in this pull request:
https://github.com/code4sac/trash-ai/pull/181

But also in many others that you will see referenced in the issues.

What is the best way to proceed with the review from here?
