### 2023-08-21 #5136 @arfon
Thanks for the update @wincowgerDEV. As I think has been discussed earlier in this thread, automated tests are desirable but not essential to pass a JOSS review, provided there are other mechanisms (e.g., testing _procedures_ that can be followed to verify the behaviour of the software.

@domna, @luxaritas – based on @wincowgerDEV's update here, if you could take one final pass on your checkboxes above.

> What is the best way to proceed with the review from here?

@wincowgerDEV - based on what feels like a converging conversation, if the reviewers are able to verify your responses and updates to the package I think we're close to being done here 😄
