### 2023-08-24 #5136 @wincowgerDEV
Hey @domna and @luxaritas, I am thankful for all the effort you have already put in and recognize that it took me a while to go through the updates so it's challenging to get started again. Below I go through each of the check boxes you have open and provide my rationale for why it should be checked:

- Substantial scholarly effort (@domna): discussed here https://github.com/openjournals/joss-reviews/issues/5136#issuecomment-1428137204
- Installation instructions (@domna): described here https://github.com/code4sac/trash-ai/blob/aws/trashai-staging/docs/localdev.md
-  Example usage (@domna): described here https://github.com/code4sac/trash-ai#demo and here https://github.com/code4sac/trash-ai/blob/aws/trashai-staging/paper.md
- Functionality documentation (@domna and @luxaritas): https://github.com/code4sac/trash-ai/tree/aws/trashai-staging/docs
- Automated tests (@domna and @luxaritas): There are automated tests we are working on but for now we just have manual tests that will validate the schema and functionality of the tool they are described in the video https://github.com/code4sac/trash-ai#demo
- Community guidelines (@domna): https://github.com/code4sac/trash-ai#contribute
- State of the field (@domna): https://github.com/code4sac/trash-ai/blob/aws/trashai-staging/paper.md#statement-of-need
- References (@domna and @luxaritas): https://github.com/code4sac/trash-ai/blob/aws/trashai-staging/paper.bib
