### 2023-08-27 #5136 @luxaritas
@wincowgerDEV Thanks for the updates. For the remaining things on my end:
- The references check looks to be good now
- For functionality documentation, would it be easy to add a basic textual guide either to the readme or a separate markdown file in the docs directory with a link from the readme? Doesn't need to be exhaustive, but a video can be harder to use as an at-a-glance overview or quick reference
- Similarly for automation, could you add maybe a checklist in a file in the docs directory that describes the testing procedure? The demo video is great, but doesn't provide a clear list of steps of things that should be checked to know the software is behaving appropriately.

I'm hoping both of those things should be very quick to write up.
