### 2023-08-28 code4sac/trash-ai#190 @wincowgerDEV
[Feature]: Review 2
### Guidelines

- [X] I agree to follow this project's Contributing Guidelines.

### Description

Reviewer @luxaritas requests

### Problem

Video documentation is difficult to follow for some users.

### Proposed Solution

- [x] For functionality documentation, add a basic textual guide (with screenshots as appropriate) either to the readme or a separate markdown file in the docs directory with a link from the readme? Doesn't need to be exhaustive.

- [x] Similarly for automation, add a checklist in a file in the docs directory that describes the testing procedure? The demo video is great, but doesn't provide a clear list of steps of things that should be checked to know the software is behaving appropriately.

### Alternatives Considered

Video documentation.
