### 2023-08-28 code4sac/trash-ai#190 @wincowgerDEV
See comment: https://github.com/openjournals/joss-reviews/issues/5136#issuecomment-1694759905
