### 2023-09-08 code4sac/trash-ai#190 @wincowgerDEV
We do have some manual testing documentation already which is also cross-referenced in the readme: https://github.com/code4sac/trash-ai/tree/production/frontend/__tests__

I will add the checklist suggested.
