### 2023-09-08 dcb2265
Update README.md
Resolving testing proceedure. https://github.com/openjournals/joss-reviews/issues/5136#issuecomment-1695915934
