### 2023-09-08 b2c4c70
Update and rename functionality to functionality.md
functionality update
