### 2023-09-08 code4sac/trash-ai#190 @wincowgerDEV
Closing with these commits:

b2c4c70a490a0b6372b25b9080ec7cb69bd35881

3a6e27257daa49e9d2ae834d2c2cd76dfaa9c653

51aed34d5ff2958b074c5e148436eb4c916c0517

dcb2265065e469013879f02de53f016a1d2aae3f
