### 2023-09-08 #5136 @wincowgerDEV
@luxaritas, I completed your additional requests liked in the issue above https://github.com/code4sac/trash-ai/issues/190. Let me know if there is anything else I can do.

@domna, have you had a chance to re-review the repo or have anything else I can do? As a reminder please see this for my rationale for acceptance: https://github.com/openjournals/joss-reviews/issues/5136#issuecomment-1692542881
