### 2023-09-11 #5136 @domna
I agree with @luxaritas point, that it would be nice to have a top-level README in the docs folder which is automatically displayed when navigating there and gives a short overview.

Despite this everything looks good to me and it's good to go from my side.
