### 2023-09-13 #5136 @wincowgerDEV
> I agree with @luxaritas point, that it would be nice to have a top-level README in the docs folder which is automatically displayed when navigating there and gives a short overview.
>
> Despite this everything looks good to me and it's good to go from my side.

Thanks @domna for your positive feedback and additional comment. I just created the top level docs as per your suggestion: https://github.com/code4sac/trash-ai/commit/b46835c2a83609de0e690ddc3e854803cce7522e

@luxaritas, is there anything additional you would like to see or are you good with checking the final boxes per my updates?

Warm Regards,
Win
