### 2023-09-13 7957f51
Review2 (#191)
* Update README.md

Resolving testing proceedure. https://github.com/openjournals/joss-reviews/issues/5136#issuecomment-1695915934

* Create functionality

* add functionality link

* Update and rename functionality to functionality.md

functionality update

* https://github.com/openjournals/joss-reviews/issues/5136#issuecomment-1713277787

Responding to @domna's reviewer comment.
