+++ docs/README.md
@@ -1,0 +1,17 @@
+# Top Level Overview of Docs
+
+## functionality.md
+Provides information about the common workflows and functions of the app. 
+
+## git-aws-account-setup.md
+Describes how you can set up git based deployment of the app to AWS. 
+
+## github-actions-deployment-role.md
+How to set up the github deployment role for deploying with github actions. 
+
+## github-actions.md
+Information about the basic github action files and deployment strategy. 
+
+## localdev.md
+Instructions on how to deploy the app locally. 
+
